refactor: extract shared length-to-px conversion into a trait
