+++ includes/resources/Design_Tokens/Projection/Adapter/Icon_Size_Adapter.php
@@ -3,6 +3,7 @@
 namespace KadenceWP\KadenceBlocks\Design_Tokens\Projection\Adapter;
 
 use KadenceWP\KadenceBlocks\Design_Tokens\Projection\Adapter\Contracts\Abstract_Adapter;
+use KadenceWP\KadenceBlocks\Design_Tokens\Projection\Traits\Converts_Length_To_Px;
 use KadenceWP\KadenceBlocks\Design_Tokens\Resolver\Token_Resolver;
 
 /**
@@ -28,6 +29,8 @@ use KadenceWP\KadenceBlocks\Design_Tokens\Resolver\Token_Resolver;
  * @since TBD
  */
 final class Icon_Size_Adapter extends Abstract_Adapter {
+
+	use Converts_Length_To_Px;
 
 	/**
 	 * @since TBD
@@ -86,27 +89,5 @@ final class Icon_Size_Adapter extends Abstract_Adapter {
 		$attributes['size'] = $px;
 
 		return $attributes;
-	}
-
-	/**
-	 * Convert a resolved CSS length to a raw pixel number, assuming the browser/CSS default root
-	 * font size (16px) for `rem`/`em` — the only units this token family's baseline values use.
-	 * Returns null for a value this adapter cannot safely convert (already px, or an unrecognized
-	 * unit), so the caller can leave the attribute as WordPress's own default rather than guess.
-	 *
-	 * @since TBD
-	 *
-	 * @param string $length A resolved CSS length, e.g. "1.5rem", "24px".
-	 *
-	 * @return float|null The pixel value, or null when the unit is not rem/em/px.
-	 */
-	private function to_px( string $length ): ?float {
-		if ( ! preg_match( '/^(-?[0-9.]+)(px|rem|em)$/', trim( $length ), $matches ) ) {
-			return null;
-		}
-
-		$number = (float) $matches[1];
-
-		return $matches[2] === 'px' ? $number : $number * 16.0;
 	}
 }
